lab6/requirements.txt
fastapi
uvicorn
pymongo
python-jose[cryptography]
python-multipart
pydantic
pytest
httpx
